pip-venv-requirements.txt
pip
jmespath==0.10.0
ansible-core==2.14.18
jinja2==3.1.2
ansible-lint
pytest
pytest-asyncio
pytest-cov
flake8
autopep8
black
pymongo
sanic>=23.0.0
motor>=3.0.0
